Import as HTTPupload
In Neos 5 this was not working anymore. I wrote the stream in a memory and gave it then to the function. I think there should be a more elegant version, but I'm right now not familiar enough with the NEOS API to provide that. Do you have a better solution?